refactor: move text::Masked to text::masked::Masked
Re-exports the Masked type at text::Masked
